tab styles: Inplace and Indented
Add support for tab styles in PrettyTabbedDoc. At the moment,
every tab generated by TabbedTokenDoc is the 'Inplace' style.

Inplace tabs are allowed to start anywhere after their parent.
This could be the same column as the parent, or a much deeper
column, e.g.:

  |the quick fox |jumped
  |              |over
  |              |something
  |              |
  ^parent tab    ^child tab (inplace)

Indented tabs are always placed with an indent relative to
their parent:

  |the quick fox
  |  |jumped
  |  |over
  |  |something
  |  |
  |  ^child tab (indented)
  ^parent tab