Tab style: indentedExactlyBy. Used to handle comments better for functions with clauses
Indent-style tabs now support optional upper bounds, in addition to the
(previously already supported) optional lower bound. This way, we can
constrain the indentation of a tab to a particular value, if necessary.

One application of this is in the formatting of functions with clauses:
```
fun foo x = ...
  | foo y = ...
  | foo z = ...
```

We can use a child tab to align the bars `|` at exactly 2 offset from
the left alignment of the function. Previouly, this was accomplished by
inserting two spaces in exactly the right place, which IMO is a bit
yuck.

But, avoiding yuck wasn't the main motivation for this patch. The purpose
here was to play nicely with inserting comments for functions with clauses.

Previously, comments between clauses would jump to the far left alignment.
Now, they align with the beginning of the clause. This is more consistent
with how comments are handled in other branch-based constructs, such as
case expressions.

**OLD OUTPUT**:
```
(* A *)
fun foo x = ...

(* B *)
  | foo y = ...

(* C *)
  | foo z = ...
```

**NEW OUTPUT**:
```
(* A *)
fun foo x = ...

  (* B *)
  | foo y = ...

  (* C *)
  | foo z = ...
```