Two small improvements to let-in-end formatting
First, multi-declaration `let`s are now rigid. The smaller collapsed form
is still allowed for single declarations.

OLD
```
val _ =
  let val x = 5 val y = 6
  in x + y
  end
```

NEW
```
val _ =
  let
    val x = 5
    val y = 6
  in
    x + y
  end
```

Second, an empty declaration that contains a comment inside a let now
looks better. This situation happens occasionally in practice if the
declaration inside a let has been commented-out, for example as shown
below.

OLD
```
val _ =
  let (* val _ =
           print "hello world" *) in
    ()
  end
```

NEW
```
val _ =
  let
  (* val _ =
       print "hello world" *)
  in ()
  end
```

It'd be nice if the comment between let-in was indented, but oh well.
One thing on the todo list at a time...